requirements.txt
beautifulsoup4==4.12.2
Django==4.2.5
g4f==0.1.4.6
openai==0.28.1
pypinyin==0.49.0
python-decouple==3.8
Requests==2.31.0
